edit too  much files

+++ archives-manage-entity/src/main/java/com/yintu/rixing/make/MakeBorrowAuditor.java
@@ -21,7 +21,7 @@ import java.util.Date;
 @Data
 @EqualsAndHashCode(callSuper = true)
 @TableName("make_borrow_auditor")
-@ApiModel(value="MakeBorrowAuditor对象", description="借阅申请文件和审核人中间表")
+@ApiModel(value = "MakeBorrowAuditor对象", description = "借阅申请文件和审核人中间表")
 public class MakeBorrowAuditor extends BaseEntity {
 
     private static final long serialVersionUID = 1L;
@@ -36,11 +36,11 @@ public class MakeBorrowAuditor extends BaseEntity {
 
     @ApiModelProperty(value = "排序")
     @TableField("sort")
-    private Integer sort;
+    private Short sort;
 
     @ApiModelProperty(value = "当前人是否在审核中 1.是 0.否")
     @TableField("activate")
-    private Integer activate;
+    private Short activate;
 
     @ApiModelProperty(value = "审核内容")
     @TableField("context")
@@ -56,11 +56,11 @@ public class MakeBorrowAuditor extends BaseEntity {
 
     @ApiModelProperty(value = "是否处理1.已处理 0.待处理")
     @TableField("is_dispose")
-    private Integer isDispose;
+    private Short isDispose;
 
     @ApiModelProperty(value = "2.审核中 3.通过 4.拒绝 5.转交")
     @TableField("audit_status")
-    private Integer auditStatus;
+    private Short auditStatus;
 
     @ApiModelProperty(value = "审核完成时间")
     @TableField("audit_finish_time")
